Add files via upload
For some reason the footer is in a funky place. I will need to work on it some more. I need to make spaces appear between some of the sections, and the transitions from one state to another is rocky. Had some severe problems getting the website to display in Microsoft Edge for some reason. I normally don't use Edge. I'm more of a firefox kind of gal, but Edge is the default browser on my laptop. It works better in Chrome.